Fixed decimal issue when operating
An issue persisted where, when an operation was applied, and a trailing
decimal was present (like '6.' or '0.'), the decimal would stick and
go over to the second num, including the digits that precede the number.
Therefore, the issue was addressed by converting the first number and
the result content to a Number() object before comparing them (to make
sure that the correct number was being displayed after clicking on
an operation).